+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Fixed issue in sra module where increment was larger than layer, then failed
 * Fixed bug where 'void_ratio' method for unit weight in BI2014 triggering added dry weight twice
 * Added support for loading cpt files that do not have the u2 column
+* Updated factor of safety colors for plotting, added color map and colors accessible as static values
 
 0.5.3 (2019-04-08)
 -------------------